stock_service/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
prometheus_client